Item Stock Update successfully

+++ Models/Item.cs
@@ -10,7 +10,7 @@ namespace sm_backend.Models
         public decimal CostOfItem { get; set; }
         public decimal RealItemCost { get; set; }
 
-        public int TotalQuantity { get; set; }
+        public decimal TotalQuantity { get; set; }
         public decimal TotalAmount { get; set; }
         public string? TypeOfItem { get; set; }
